controller y service actualizado

+++ proyecto-gestion-solicitudes-main/src/main/java/com/campus360/solicitudes/Client/CatalogoClient.java
@@ -19,7 +19,7 @@ public class CatalogoClient {
         this.catalogoBaseUrl = catalogoBaseUrl;
     }
 
-    public ServicioInfoResponse obtenerServicio(Long catalogoId) {
+    public ServicioInfoResponse obtenerServicio(int catalogoId) {
 
         String url = catalogoBaseUrl + "/servicios/" + catalogoId;
